Begin renaming the argocd environment by copying it
The environment creates a helm chart application named argocd, so if
someone reference our environment in a tanka application object, and
they name it argocd, there'll be a naming conflict which could cause
argocd to uninstall.

It would be better to have our environment name it something less likely
to conflict, e.g. `argocd-helm-chart`, but it already exists in every
vcluster as `argocd`, so that would be an expensive name change.