Pull cluster.json from running container instead of secret
We opted to move the secret into the host cluster (so that they could be
managed automatically with git ops), which means the secret is no longer
accessible directly to the vcluster.

However, if a vcluster pod tries to mount `cluster-details`, then a pod
in the host cluster will try to mount `cluster-details-x-etc-x-vcluster`
so as long as that exists, this will work just fine. It will even work
if someone decides to eject.