Upgrade Prometheus and support remote_write to Mimir
We extend the pattern of conditionally including the secret mount if
alert managers or remote write targets are defined in cluster.json.
There is probably a better, unconditional solution, but it will require
automation around generating the certificates/secrets per vcluster.

We also upgrade Prometheus and the Push Gateway to the versions used
on the traditional servers; no config changes or migrations are required.